Create files for each class and import them to main.js
