+++ index.html
@@ -36,6 +36,6 @@
         <button type="submit">Add book</button>
       </form>
     </main>
-    <script src="assets/js/main.js" defer></script>
+    <script type="module" src="assets/js/main.js" defer></script>
   </body>
 </html>
